refactor: remove duplicate elements and Obejct, class
