+++ src/main/java/com/dope/breaking/Security/SecurityController/Oauth2LoginController.java
@@ -4,17 +4,14 @@ package com.dope.breaking.Security.SecurityController;
 import com.dope.breaking.Security.Jwt.JwtTokenProvider;
 import com.dope.breaking.Security.UserInfoDto.UserDto;
 import com.dope.breaking.domain.user.User;
-import com.dope.breaking.Security.UserInfoDto.UserService;
+import com.dope.breaking.service.UserService;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.http.*;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
