--- src/main/java/com/dope/breaking/Security/UserInfoDto/UserRepository.java
@@ -1,11 +1,0 @@
-package com.dope.breaking.Security.UserInfoDto;
-
-import com.dope.breaking.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-
-}
