+++ src/main/java/com/dope/breaking/repository/UserRepository.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User,Long> {
 
+    Optional<User> findByUsername(String username);
     Optional<User> findByNickname(String nickname);
     Optional<User> findByPhoneNumber(String phoneNumber);
     Optional<User> findByEmail(String email);
