+++ src/main/java/com/dope/breaking/service/UserService.java
@@ -1,9 +1,21 @@
 package com.dope.breaking.service;
 
+import com.dope.breaking.domain.user.User;
+import com.dope.breaking.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+@Slf4j
+@RequiredArgsConstructor
+@Service
 public class UserService {
+
+    private final UserRepository userRepository;
 
     public static boolean isValidEmail(String email) {
 
@@ -20,4 +32,9 @@ public class UserService {
         }
         return err;
     }
+
+    public Optional<User> findbyUsername(String username) {
+        return userRepository.findByUsername(username);
+    }
+
 }
